adjust to changes in GAP
This pull request belongs to gap-system/gap/pull/5395:

When `NewCompanionMatrix`, `NewIdentityMatrix`, `NewMatrix`, `NewVector`,
`NewZeroMatrix`, `NewZeroVector` are changed from constructors to
tag based operations, the method installations in the cvec package
must be adjusted.

The current proposal makes the cvec code compatible with the old and
the new behaviour of the operations.

(Additionally, use `NewVector` not `NewRowVector`;
the obsolescent synonym `NewRowVector` is still available in GAP
but using only `NewVector` makes it easier to find out where the
operation is used.)